+++ shuup/admin/modules/sales_units/views/__init__.py
@@ -6,7 +6,12 @@
 # This source code is licensed under the OSL-3.0 license found in the
 # LICENSE file in the root directory of this source tree.
 
-from .edit import SalesUnitEditView
-from .list import SalesUnitListView
+from .edit import DisplayUnitEditView, SalesUnitEditView
+from .list import DisplayUnitListView, SalesUnitListView
 
-__all__ = ["SalesUnitEditView", "SalesUnitListView"]
+__all__ = [
+    "DisplayUnitEditView",
+    "DisplayUnitListView",
+    "SalesUnitEditView",
+    "SalesUnitListView",
+]