+++ shuup/admin/modules/sales_units/views/edit.py
@@ -9,7 +9,7 @@
 from __future__ import unicode_literals
 
 from shuup.admin.utils.views import CreateOrUpdateView
-from shuup.core.models import SalesUnit
+from shuup.core.models import DisplayUnit, SalesUnit
 from shuup.utils.multilanguage_model_form import MultiLanguageModelForm
 
 
@@ -24,3 +24,16 @@ class SalesUnitEditView(CreateOrUpdateView):
     form_class = SalesUnitForm
     template_name = "shuup/admin/sales_units/edit.jinja"
     context_object_name = "sales_unit"
+
+
+class DisplayUnitForm(MultiLanguageModelForm):
+    class Meta:
+        model = DisplayUnit
+        exclude = ()  # All the fields!
+
+
+class DisplayUnitEditView(CreateOrUpdateView):
+    model = DisplayUnit
+    form_class = DisplayUnitForm
+    template_name = "shuup/admin/sales_units/edit_display_unit.jinja"
+    context_object_name = "display_unit"